requirements (1).txt
# DermAI Backend — Requirements
fastapi==0.111.0
uvicorn==0.29.0
sqlalchemy==2.0.30
bcrypt==4.1.3
python-jose==3.3.0
razorpay==1.4.1
httpx==0.27.0
python-multipart==0.0.9
pydantic==2.7.1
groq==0.9.0
optimum[onnxruntime]==1.19.1
onnxruntime==1.18.0
transformers==4.41.0
psycopg2-binary==2.9.9
